LPD-58579 Packages rename
